Went to /W4, wrapped GLM includes to get rid of C4021, nuked warnings out of own source

+++ UnnamedEngine/Sources/Engine/Base/Managers/EntityAdmin.h
@@ -43,7 +43,7 @@ private:
 	T* GetComponent(uint32_t entityID)
 	{
 		const ComponentFlag flag = ComponentGroup<T>();
-		void* componentPtr = mComponentPools[flag]->GetComponent(entityID);
+		void* componentPtr = mComponentPools[static_cast<int>(flag)]->GetComponent(entityID);
 		T* component = static_cast<T*>(componentPtr);
 		return(component);
 	}
